Changed hour columns to floats

+++ db/migrate/20130524022627_create_service_orders.rb
@@ -5,7 +5,7 @@ class CreateServiceOrders < ActiveRecord::Migration
       t.text :description
       t.integer :category_id
       t.integer :user_id
-      t.integer :budget, default: 0
+      t.float :budget, default: 0
       t.boolean :completed, default: false
 
       t.timestamps
