+++ db/migrate/20130524022938_create_confirmations.rb
@@ -3,10 +3,10 @@ class CreateConfirmations < ActiveRecord::Migration
     create_table :confirmations do |t|
       t.integer :service_order_id
       t.date :performed_on
-      t.integer :travel_hours, default: 0
-      t.integer :straight_hours, default: 0
-      t.integer :overtime_hours, default: 0
-      t.integer :doubletime_hours, default: 0
+      t.float :travel_hours, default: 0
+      t.float :straight_hours, default: 0
+      t.float :overtime_hours, default: 0
+      t.float :doubletime_hours, default: 0
       t.text :description
       t.string :number
       t.integer :user_id
